Refactor logic into new scrollUtils.ts
